requirements.txt
streamlit
requests
pillow
numpy
tensorflow
keras
scikit-learn
matplotlib
opencv-python